requirements.txt
APScheduler==3.5.3
asn1crypto==0.24.0
bcrypt==3.1.4
cffi==1.11.5
cryptography==2.3.1
gevent==1.3.6
greenlet==0.4.15
idna==2.7
netmiko==2.2.2
paramiko==2.4.2
pyasn1==0.4.4
pycparser==2.19
PyNaCl==1.3.0
pyserial==3.4
pytz==2018.7
PyYAML==3.13
scp==0.11.0
six==1.11.0
textfsm==0.4.1
tzlocal==1.5.1
